Finishes Quizzer.
The Quizzer should now be finished. Please try to break it or look for
flaws / give me feedback about what you hate / like. It will save to the
db as well... Only concern is that the results model's username field
should probably be a userid field to relate to the user's id.